+++ components/service/block_storage/block_store/device/file/file_block_store.c
@@ -162,7 +162,7 @@ static psa_status_t file_block_store_close(void *context, uint32_t client_id,
 }
 
 static psa_status_t file_block_store_read(void *context, uint32_t client_id,
-					  storage_partition_handle_t handle, uint32_t lba,
+					  storage_partition_handle_t handle, uint64_t lba,
 					  size_t offset, size_t buffer_size, uint8_t *buffer,
 					  size_t *data_len)
 {
@@ -199,7 +199,7 @@ static psa_status_t file_block_store_read(void *context, uint32_t client_id,
 }
 
 static psa_status_t file_block_store_write(void *context, uint32_t client_id,
-					   storage_partition_handle_t handle, uint32_t lba,
+					   storage_partition_handle_t handle, uint64_t lba,
 					   size_t offset, const uint8_t *data, size_t data_len,
 					   size_t *num_written)
 {
@@ -237,7 +237,7 @@ static psa_status_t file_block_store_write(void *context, uint32_t client_id,
 }
 
 static psa_status_t file_block_store_erase(void *context, uint32_t client_id,
-					   storage_partition_handle_t handle, uint32_t begin_lba,
+					   storage_partition_handle_t handle, uint64_t begin_lba,
 					   size_t num_blocks)
 {
 	struct file_block_store *this_instance = (struct file_block_store *)context;
