+++ components/service/block_storage/block_store/device/null/null_block_store.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Arm Limited. All rights reserved.
+ * Copyright (c) 2022-2023, Arm Limited. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  *
@@ -39,7 +39,7 @@ static psa_status_t null_block_store_close(void *context,
 static psa_status_t null_block_store_read(void *context,
 	uint32_t client_id,
 	storage_partition_handle_t handle,
-	uint32_t lba,
+	uint64_t lba,
 	size_t offset,
 	size_t buffer_size,
 	uint8_t *buffer,
@@ -78,7 +78,7 @@ static psa_status_t null_block_store_read(void *context,
 static psa_status_t null_block_store_write(void *context,
 	uint32_t client_id,
 	storage_partition_handle_t handle,
-	uint32_t lba,
+	uint64_t lba,
 	size_t offset,
 	const uint8_t *data,
 	size_t data_len,
@@ -116,7 +116,7 @@ static psa_status_t null_block_store_write(void *context,
 static psa_status_t null_block_store_erase(void *context,
 	uint32_t client_id,
 	storage_partition_handle_t handle,
-	uint32_t begin_lba,
+	uint64_t begin_lba,
 	size_t num_blocks)
 {
 	struct null_block_store *null_block_store = (struct null_block_store*)context;
